+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 * The output of the keywords 'X-ADD-KEY' and 'X-DELETE-KEY' now also show the "oneline"-format to represent keys (which includes fingerprint, primary email-address, date of generation and possible expiry). (#295)
 * In the response to 'X-ADD-KEY', differentiate between 'newly imported' and 'updated' keys.
+* Parse keywords up to the first line detected as mail content, this addresses a first part of #249.
 
 
 ## [3.2.3] / 2018-05-14